Fixed: Break up admin.py
- Moves each model admin to its own file
- Creates a shared base_admin file for the base model admin

Fixes #110
