+++ django/cantusdb_project/main_app/admin/__init__.py
@@ -1,0 +1,12 @@
+from main_app.admin.century import CenturyAdmin
+from main_app.admin.chant import ChantAdmin
+from main_app.admin.differentia import DifferentiaAdmin
+from main_app.admin.feast import FeastAdmin
+from main_app.admin.genre import GenreAdmin
+from main_app.admin.notation import NotationAdmin
+from main_app.admin.office import OfficeAdmin
+from main_app.admin.provenance import ProvenanceAdmin
+from main_app.admin.rism_siglum import RismSiglumAdmin
+from main_app.admin.segment import SegmentAdmin
+from main_app.admin.sequence import SequenceAdmin
+from main_app.admin.source import SourceAdmin
