+++ django/cantusdb_project/main_app/admin/institution.py
@@ -1,0 +1,5 @@
+from django.contrib import admin
+
+
+class InstitutionAdmin(admin.ModelAdmin):
+    pass
