+++ django/cantusdb_project/main_app/admin/institution_identifier.py
@@ -1,0 +1,5 @@
+from django.contrib import admin
+
+
+class InstitutionIdentifierAdmin(admin.ModelAdmin):
+    pass
