+++ django/cantusdb_project/main_app/admin/segment.py
@@ -1,0 +1,11 @@
+from django.contrib import admin
+
+from main_app.admin.base_admin import BaseModelAdmin
+from main_app.forms import AdminSegmentForm
+from main_app.models import Segment
+
+
+@admin.register(Segment)
+class SegmentAdmin(BaseModelAdmin):
+    search_fields = ("name",)
+    form = AdminSegmentForm
